flatten param generator
